pyproject.toml
[tool.poetry]
name = "dotflow-server"
version = "0.1.0"
description = ""
authors = ["FernandoCelmer <email@fernandocelmer.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pyzmq = "^26.2.1"
jinja2 = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
